useEffect does nothing
