+++ hooks/useMap.tsx
@@ -11,43 +11,44 @@ export default function useMap(
 ) {
   const [lng, setLng] = useState(12.37);
   const [lat, setLat] = useState(51.34);
-  let userLocation: Coord = [];
+  const [userLocation, setUserLocation] = useState<Coord>();
 
   useEffect(() => {
     //check, if map actually exists
     if (!map.current) return;
+    navigator.geolocation.getCurrentPosition((position) => {
+      const userCoordinates = [
+        position.coords.longitude,
+        position.coords.latitude,
+      ];
+      //feed user location infomation to the map
+      (map.current as mapboxgl.Map).addSource("user-coordinates", {
+        type: "geojson",
+        data: {
+          type: "Feature",
+          geometry: {
+            type: "Point",
+            coordinates: userCoordinates,
+          },
+        } as any,
+      });
 
-    // get userlocation
-    userLocation = getUserLocation(map);
+      //display user location  on map
+      (map.current as mapboxgl.Map).addLayer({
+        id: "user-coordinates",
+        source: "user-coordinates",
+        type: "circle",
+      });
 
-    //   //feed user location infomation to the map
-    //   (map.current as mapboxgl.Map).addSource("user-coordinates", {
-    //     type: "geojson",
-    //     data: {
-    //       type: "Feature",
-    //       geometry: {
-    //         type: "Point",
-    //         coordinates: userLocation,
-    //       },
-    //     } as any,
-    //   });
+      //center map on user location
+      (map.current as mapboxgl.Map).flyTo({
+        center: userCoordinates as LngLatLike,
+        zoom: 14,
+      });
 
-    //   //display user location  on map
-    //   (map.current as mapboxgl.Map).addLayer({
-    //     id: "user-coordinates",
-    //     source: "user-coordinates",
-    //     type: "circle",
-    //   });
-
-    //   //center map on user location
-    //   (map.current as mapboxgl.Map).flyTo({
-    //     center: userLocation as LngLatLike,
-    //     zoom: 14,
-    //   });
-
-    //   //store user location
-    // });
-
+      //store user location
+      setUserLocation(userCoordinates);
+    });
     map.current.on("load", () => {
       // create button for centering the map manually on user
       const geolocate = new mapboxgl.GeolocateControl({
